requirements.txt
fastapi==0.63.0
Pillow==8.1.0
tensorflow==2.4.0
numpy==1.19.5
uvicorn==0.13.3
aiofiles==0.6.0
python-multipart==0.0.5
